requirements.txt
alembic==1.17.1
annotated-types==0.7.0
anyio==4.11.0
certifi==2025.10.5
charset-normalizer==3.4.4
exceptiongroup==1.3.0
greenlet==3.2.4
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.11
Mako==1.3.10
MarkupSafe==3.0.3
pydantic==2.12.3
pydantic_core==2.41.4
python-dotenv==1.2.1
python-telegram-bot==22.5
requests==2.32.5
sniffio==1.3.1
SQLAlchemy==2.0.44
tomli==2.3.0
typing-inspection==0.4.2
typing_extensions==4.15.0
urllib3==2.5.0
